Extract HF access token acquirement common code
